Move SttState from ui.home to io.input
