+++ app/src/androidTest/kotlin/org/stypox/dicio/screenshot/FakeSttInputDeviceWrapper.kt
@@ -3,7 +3,7 @@ package org.stypox.dicio.screenshot
 import kotlinx.coroutines.flow.MutableStateFlow
 import org.stypox.dicio.di.SttInputDeviceWrapper
 import org.stypox.dicio.io.input.InputEvent
-import org.stypox.dicio.ui.home.SttState
+import org.stypox.dicio.io.input.SttState
 
 class FakeSttInputDeviceWrapper : SttInputDeviceWrapper {
     override val uiState: MutableStateFlow<SttState> = MutableStateFlow(SttState.NotInitialized)
