+++ app/src/main/kotlin/org/stypox/dicio/io/input/SttInputDevice.kt
@@ -1,7 +1,6 @@
 package org.stypox.dicio.io.input
 
 import kotlinx.coroutines.flow.StateFlow
-import org.stypox.dicio.ui.home.SttState
 
 interface SttInputDevice {
     val uiState: StateFlow<SttState>
