+++ app/src/main/kotlin/org/stypox/dicio/ui/home/SttButton.kt
@@ -38,20 +38,21 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.core.content.ContextCompat
 import org.stypox.dicio.R
-import org.stypox.dicio.ui.home.SttState.Downloaded
-import org.stypox.dicio.ui.home.SttState.Downloading
-import org.stypox.dicio.ui.home.SttState.ErrorDownloading
-import org.stypox.dicio.ui.home.SttState.ErrorLoading
-import org.stypox.dicio.ui.home.SttState.ErrorUnzipping
-import org.stypox.dicio.ui.home.SttState.Listening
-import org.stypox.dicio.ui.home.SttState.Loaded
-import org.stypox.dicio.ui.home.SttState.Loading
-import org.stypox.dicio.ui.home.SttState.NoMicrophonePermission
-import org.stypox.dicio.ui.home.SttState.NotAvailable
-import org.stypox.dicio.ui.home.SttState.NotDownloaded
-import org.stypox.dicio.ui.home.SttState.NotInitialized
-import org.stypox.dicio.ui.home.SttState.NotLoaded
-import org.stypox.dicio.ui.home.SttState.Unzipping
+import org.stypox.dicio.io.input.SttState
+import org.stypox.dicio.io.input.SttState.Downloaded
+import org.stypox.dicio.io.input.SttState.Downloading
+import org.stypox.dicio.io.input.SttState.ErrorDownloading
+import org.stypox.dicio.io.input.SttState.ErrorLoading
+import org.stypox.dicio.io.input.SttState.ErrorUnzipping
+import org.stypox.dicio.io.input.SttState.Listening
+import org.stypox.dicio.io.input.SttState.Loaded
+import org.stypox.dicio.io.input.SttState.Loading
+import org.stypox.dicio.io.input.SttState.NoMicrophonePermission
+import org.stypox.dicio.io.input.SttState.NotAvailable
+import org.stypox.dicio.io.input.SttState.NotDownloaded
+import org.stypox.dicio.io.input.SttState.NotInitialized
+import org.stypox.dicio.io.input.SttState.NotLoaded
+import org.stypox.dicio.io.input.SttState.Unzipping
 import org.stypox.dicio.ui.theme.AppTheme
 import org.stypox.dicio.ui.util.LoadingProgress
 import org.stypox.dicio.ui.util.SmallCircularProgressIndicator
